# swagger.yaml
swagger: "2.0"
info:
  title: Space Mission Control API
  version: 1.0.1
  description: |
    API for managing space missions, control users, astronauts, and launch vehicles.
    All error strings are required to be clear descriptive message. The descriptive message won't change the marks you receive when being automarked as long as you make an attempt for each one to be custom, but they will affect your code quality mark.
schemes:
  - http
...
components:
  prim:
    ControlUserSessionId:
      type: string
      description: The sessionID representing a logged in controlUser
      example: "abcd1234"
    Error:
      type: string
      example: "Meaningful error message"
...
    MissionId:
      type: number
      example: 1
...
    LaunchId:
      type: number
      example: 4
...
    LaunchVehicleId:
      type: number
      example: 1
...
    Timestamp:
      type: number
      example: 1739912980
      description: In seconds. You can find more information that here https://en.wikipedia.org/wiki/Unix_time
    LaunchSummary:
      type: string
      description: String representation of the launch in the format [mission.target] mission.name - launch.createdAt 
      example: "[Earth orbit] Mercury - 1192323092"
...
    LaunchCalculationParameters:
      type: object
      properties:
        targetDistance:
          type: number
          description: The distance to the target for this launch in metres
          example: 12000
        fuelBurnRate:
          type: number
          description: The amount of thrust fuel that will be consumed per second (kg/s)
          example: 20
        thrustFuel:
          type: number
          description: The amount of thrust fuel (kg) that will be added to the launch vehicle at the start of the launch
          example: 1000
        activeGravityForce:
          type: number
          format: float
          description: The force of gravity at the launch point (m/s)
          example: 9.8
        maneuveringDelay:
          type: number
          format: int32
          description: The number of seconds to wait before firing maneuvering jets once you enter a MANEUVERING state.
          example: 2
...
  group:
...
    LaunchVehicle:
      type: object
      properties:
        launchVehicleId:
          $ref: '#/components/prim/LaunchVehicleId'
        name:
          type: string
          description: Name of the launch vehicle
          example: Saturn V
        timeAdded:
          $ref: '#/components/prim/Timestamp'
        timeLastEdited:
          $ref: '#/components/prim/Timestamp'
        maxCrewWeight:
          type: number
          example: 500
        maxPayloadWeight:
          type: number
          example: 700
        launchVehicleWeight:
          type: number
          example: 4000
        thrustCapacity:
          type: number
          example: 1000000
        startingManeuveringFuel: 
          type: number
          example: 10
        retired:
          type: boolean
          example: false
        launches:
          type: array
          items:
            type: object
            properties:
              launch:
                $ref: '#/components/prim/LaunchSummary'
              state:
                type: string
                example: MISSION_COMPLETE
                description: The current state of the launch.
...
  path:
    MissionId:
      in: path
      name: missionid
      required: true
      schema:
        $ref: '#/components/prim/MissionId'
...
    LaunchVehicleId:
      in: path
      name: launchvehicleid
      required: true
      schema:
        $ref: '#/components/prim/LaunchVehicleId'
...
  header:
    ControlUserSessionId:
      in: header
      name: controlUserSessionId
      required: true
      schema:
        $ref: '#/components/prim/ControlUserSessionId'
  body:
...
    LaunchVehicleCreate:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          name:
            type: string
            description: The name of this launch vehicle
            example: Saturn V
          description:
            type: string
            description: A short description of this launch vehicle
            example: Multi-stage booster rocket
          maxCrewWeight:
            type: number
            description: The maximum allowed combined weight (kg) of astronauts allocated to this launch vehicle.
            example: 500
            format: int32
          maxPayloadWeight:
            type: number
            description: The maximum allowed weight (kg) of the payload that this launch vehicle can accommodate.
            example: 1000
            format: int32
          launchVehicleWeight:
            type: number
            description: The weight (kg) of the the basic launch vehicle without any payload or astronauts.
            example: 4000
            format: int32
          thrustCapacity:
            type: number
            description: The thrust capacity (N) that this launch vehicle is capable of.
            example: 1000000
            format: int32
          maneuveringFuel:
            type: number
            description: The amount of maneuvering fuel in units that this launch vehicle always starts with.
            example: 10
            format: int32
...
    PayloadCreate:
      type: object
      properties:
        description:
          type: string
          description: A description of the payload that is being sent up on this space mission
          example: UNSW Cubesat
        weight: 
          type: number
          description: The weight of the payload in kilograms
          example: 400
    LaunchCreate:
      in: body
      name: body
      required: true
      schema:
        type: object
        properties:
          launchVehicleId:
            $ref: '#/components/prim/LaunchVehicleId'
          payload:
            $ref: '#/components/body/PayloadCreate'
          launchParameters:
            $ref: '#/components/prim/LaunchCalculationParameters'
...
  return:
...
    Error:
      properties:
        error:
          $ref: '#/components/prim/Error'
    Empty:
      type: object
...
    LaunchVehicleNew:
      type: object
      properties:
        launchVehicleId:
          $ref: '#/components/prim/LaunchVehicleId'
...
    LaunchNew:
      type: object
      properties:
        launchId:
          $ref: '#/components/prim/LaunchId'
...
paths:
...
  /v1/admin/launchvehicle:
    post:
      summary: Add a new launch vehicle to our available launch vehicles 
      description: Given the basic details, create a new launch vehicle. A launch vehicle does not belong to any specific controlUser.
      tags:
        # - "By Topic: Astronaut Management"
        - "Iteration 3 (Example Adaptation targets)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/body/LaunchVehicleCreate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/LaunchVehicleNew'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - Name contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - Name is less than 2 characters or more than 20 characters
            - Description contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - Description is less than 2 characters or more than 50 characters
            - maximumCrewWeight < 100 or > 1000
            - maximumPayloadWeight < 100 or > 1000
            - launchVehicleWeight < 1000 or > 100000
            - thrustCapacity < 100000 or > 10000000
            - maneuveringFuel < 10 or > 100 
          schema:
            $ref: '#/components/return/Error'
  /v1/admin/launchvehicle/{launchvehicleid}:
    delete:
      summary: Retire a launch vehicle
      description: |
       - Given a particular launch vehicle, remove it from the list of available launch vehicles by setting its retired status to true.
      tags:
        # - "By Topic: Astronaut Management"
        - "Iteration 3 (New)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/LaunchVehicleId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - launchvehicleid is invalid
            - The Launch Vehicle is currently assigned to an active launch (a launch which is not in the state: ON_EARTH)
          schema:
            $ref: '#/components/return/Error'

    get:
      summary: Get info about a launch vehicle.
      description: Get all of the relevant information about the current launch vehicle including launch history.
      tags:
        # - "By Topic: Astronaut Management"
        - "Iteration 3 (Example Adaptation targets)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/LaunchVehicleId'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/group/LaunchVehicle'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
           Any of the following errors:
           - launchvehicleid is invalid
    put:
      summary: Edit an existing launch vehicle.
      description: Given the new basic details, edit the details of that launch vehicle.
      tags:
        # - "By Topic: Astronaut Management"
        - "Iteration 3 (New)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/LaunchVehicleId'
        - $ref: '#/components/body/LaunchVehicleCreate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/Empty'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        400:
          description: |
            If any of the following are true:
            - launchvehicleid is invalid
            - The Launch Vehicle is currently allocated to an active launch (a launch which is not in the state: ON_EARTH).
            - Name contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - Name is less than 2 characters or more than 20 characters
            - Description contains characters other than lowercase letters, uppercase letters, spaces, hyphens, or apostrophes
            - Description is less than 2 characters or more than 50 characters
            - maximumCrewWeight < 100 or > 1000
            - maximumPayloadWeight < 100 or > 1000
            - launchVehicleWeight < 1000 or > 100000
            - thrustCapacity < 100000 or > 10000000
            - maneuveringFuel < 10 or > 100 
          schema:
            $ref: '#/components/return/Error'
...
  /v1/admin/mission/{missionid}/launch:
    post:
      summary: Organise a launch based on a space mission
      description: Given the basic details, create a new launch. A launch is associated to a particular space mission and the specific controlUser who is in charge of that mission.
      tags:
        # - "By Topic: Astronaut Management"
        - "Iteration 3 (Example Adaptation targets)"
      parameters:
        - $ref: '#/components/header/ControlUserSessionId'
        - $ref: '#/components/path/MissionId'
        - $ref: '#/components/body/LaunchCreate'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/components/return/LaunchNew'
        401:
          description: ControlUserSessionId is empty or invalid (does not refer to valid logged in user session)
          schema:
            $ref: '#/components/return/Error'
        403:
          description: MissionId is empty, invalid or not associated with the current controlUser
        400:
          description: |
            If any of the following are true:
            - launchvehicleid is invalid
            - launchvehicleid is currently in another active launch (a launch which is not in the state: ON_EARTH).
            - launchvehicleid refers to a Launch Vehicle that is 'retired'
            - Payload description is empty or greater than 400 characters
            - Payload weight is greater than the maximum payload weight for the launch vehicle
            - Any LaunchCalculationParameters is < 0
            - manueveringDelay is < 1
            - fuelBurnRate > thrustFuel
            - An initial calculation with these LaunchCalculationParameters are invalid.
            
          schema:
            $ref: '#/components/return/Error'
...